Resolve issues with JSX.Element
Vue 2.7.14 enabled jsx by default, this unfortunately introduces a conflict with the types provided through our tooling and the custom Element used in `installer-win32-gen.tsx`. The most straightforward path that I could devise to resolve this is to remove the custom Element class and just use the JSX.Element that is bundled with our new tooling.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>